rename file of step from step_plan to step
