+++ crates/types/src/lib.rs
@@ -67,7 +67,7 @@ pub mod sensor_data;
 pub mod sole_pressure;
 pub mod sonar_obstacle;
 pub mod sonar_values;
-pub mod step_plan;
+pub mod step;
 pub mod support_foot;
 pub mod walk_command;
 pub mod whistle;
